check-and-install newer CLI version when available (#33)
If available, prompts the user to install the newer version.
User may choose to install or cancel; either way, the user's original command then continues.

To run:
1. Grab an old version of the styra CLI executable. If you do not have
one handy, create one by:
(a) edit `STYRA_CLI_RELEASE_VERSION` in the fetchdb root-level Makefile
to a lesser semantic version number.
(b) rebuild the `styra` binary with `make styra-cli-bin`
(c) replace your active `styra` binary with this newly old copy (likely
`cp ./build/styra /usr/local/bin`).

2. Execute any styra link command from the VSCode command palette. Makes
no difference which command as each one does the same startup checks.
Observe the version prompt.

The first time through, the system will do an update check, because it
will notice it has never been done before.
But once it records that, it does future checks on a periodic basis,
governed by the VSCode setting `Styra > Check Update Interval`. That
gives the user control over how often a check-for-update is
done—particularly relevant for exercising this code! The default is 1
day. Change it to 0, though, if you want to force a check, when you run
every command. (That forces the check, not the update; the update only
occurs if there is a newer version.)

STY-15367