introduce executor to manage command flow (#37)
### :nut_and_bolt: What code changed, and why?

Introduce command executor to wrap individual command invocation and
automatically apply the necessary leading and trailing steps, rather
than relying on code convention that the developer is supposed to
employ.

This greatly simplifies what a basic command looks like. Example: here
is the revised `run` method for the `Styra Link: Test` command:
```
  async run(): Promise<ReturnValue> {
    const result = await new CommandRunner().runStyraCmd(['link', 'test']);
    info(result);
    return ReturnValue.Completed;
  }
```

It also simplifies the messages emitted to the user for progress and
particularly for exceptions. Now the executor is essentially the main
arbiter for this.

The new tests:
```
  Executor
    ✓ command CONTINUES when checkStartup returns true (1 ms)
    ✓ command ABORTS when checkStartup returns false
    ✓ reports correct info to user when command completes
    ✓ reports correct info to user when command is terminated by user
    ✓ reports correct info to user when command fails
```

STY-15609