remove `path` from vscode settings (#41)
The previous use of the `path` settting here was slightly bogus. It
checked for an installed file existing there, but it did not install or
update to that location. Either need to fix it or remove it. Though not
a huge effort, it is a distraction from MVP, so this PR removes it.

STY-15615