prompt user to create a tenant if needed (#40)
Previously the startup checks for any Styra Link command in VSCode would
check for the presence of the DAS configuration file (~/.styra/config)
and, if not found, would prompt the user to enter the requisite URL and
token values.
This PR modifies the flow to improve the UX. If not found, the system
asks if the user has a tenant. If so, it then asks for the URL and
token. If not, it offers to take the user to signup.styra.com to get
one.

Note that if the user does not have a tenant when they invoke a Styra
Link command, whether they choose to go to get one or not, the Styra
Link command terminates. They would just need to re-invoke the command
after they get their tenant.

STY-15550