example: implement yaml2json converter
Writing the body of an example will always be 5 to 15 lines or so.
It is cumbersome to have to free-write that in JSON,
with each line surrounded by quotes,
and with any embedded quotes escaped.
Instead, let's write it in YAML which suffers neither of those.
There are some gotchas in YAML, too, but
typically much less of a concern,
notably trailing colons and leading octothorps.
Reference: https://docs.ansible.com/ansible/latest/reference_appendices/YAMLSyntax.html#gotchas

STY-15737