Add windows installer support (#48)
While the VSCode plugin could already ostensibly install the Styra CLI,
putting it on the search path after installation and recognizing whether
the Styra CLI is available to use, were not handled yet. This PR covers
both of those challenging issues.

#### Recognizing whether the Styra CLI is available

It turns out it was more straightforward to update the user's search
path persistently than it was to get it updated for immediate use.
Typically when one installs a Windows program and updates the user's
search path, you then directly invoke that new program. But here, we are
inside VSCode, installing a new program, and updating the search path.
The problem is there is no way to refresh the user's environment to
recognize that the search path has been updated! Things would be fine if
we were, say, to tell the user... "thanks for installing the CLI, now
close and re-open VSCode so things will work properly."... but then is
poor UX. Instead I modified the CLI-installation-checker to first see if
it can access the Styra CLI. If not, it does a secondary check for "in
the moment": has the CLI been installed and is it on the user's path? If
so, it is also considered "installed", so the install check passes.

Note that it was just the install-check that was at issue. Actually
using the CLI works (and worked) fine because it is run by spawning a
new child process that does pick up the updated search path
automatically.

#### Putting it on the search path

It took quite a bit of digging sifting through registry hacks and the
like until I finally uncovered the simple PowerShell commands to get and
set the user-level portion of the PATH (and being user-level, it does
_not_ require elevated privilege). Then I had to train the VSCode plugin
to invoke PowerShell commands and send data back and forth.
Turns out the same mechanism I have for spawning Styra CLI commands
could easily be repurposed to do PowerShell commands, too. And it can do
so on both Mac and Windows, though I am only using it on Windows here.

### Definition of done

- [x] Plugin has learned how to execute PowerShell commands (can
actually do this now on Mac or Windows, but only making use of it on
Windows).
- [x] Windows: Installing the CLI properly puts it on the user's search
path persistently.
- [x] Windows: When the CLI is first installed, the CLI install check
will recognize the installation.
- [x] Windows: When the CLI is installed, the CLI install check will
properly find the CLI as available on subsequent starts.
- [x] Bonus: Snippet installer needed to check whether link was
initialized before attempting to install snippets at startup.
- [x] Bonus: Link commands need to abort if the CLI cannot be accessed.

STY-15696