readme: suppress link navigation
When github.com renders an image in a readme.md like this...

```
![][http://example.com/path/image]
```

...it automatically links the image in an `<a>` tag.

```
<a href="https://camo.githubusercontent.com/8a545f12e16ff12fd...." target="_blank"><img src="https://camo.githubusercontent.com/8a545f12e16ff12f..." alt="" data-canonical-src="http://example.com/path/image" style="max-width:100%;"></a>
```

So in the case of badges, it opens the badge image in a separate tab,
which is not desirable.
There is no way to avoid the link completely; this commit just keeps the
user on the same page.

Reference: https://stackoverflow.com/a/48744957

STY-16136