Rely on OPA extension rather than re-declare the rego language (#69)
Rather than act as an alternative to the existing OPA extension, Styra
can build on top of the OPA extension, allowing us to make use of things
such as the rego language definition, the OPA binary location, root
paths, and more. The goal is to make the experience of authoring rego
with VS Code as cohesive as possible.