Add Enterprise OPA Preview API support to the VS Code Styra extension (#70)
This adds a slew of new functionality to the VS-Code Styra extension
that works with the new Enterprise OPA preview API.

This requires an Enterprise OPA instance to run in server mode with the
preview API activated. The instance must be available to accept requests
from you local machine, such as running locally, or via a tool like
`kubectl port-forward`. When these things are true, this addition allows
VS Code users to make preview requests of the Enterprise OPA instance
based on rego files they are authoring locally, within the context of
the Enterprise OPA environment.

To run a preview, the command palette (`cmd + p`) exposes several
'Enterprise OPA:' commands including Preview, Preview Package, and
Preview Selection.

In addition to the command palette, this addition exposes these commands
in the contextual menu in the editor when editing a Rego file. The user
can right-click and then run a preview.

Finally, when the CodeLens is enabled (which it is by default) Rego
files will have links rendered at the top which allow you to run a
preview with a single click.

* Preview: Previews the default location if set in settings, otherwise
falls back to running Preview Package
* Preview Package: Preview the package of the current file, sending all
Rego and data files in the defined roots to the preview API.
* Preview Selection: Run a query for the selected Rego code, sending all
Rego and data files in the defined roots to the preview API.

Configuration is offered allowing different request parameters
controlling the precise preview behavior and return. It also allows
setting a strategy so preview requests are made with just the active
file, or with all Rego files and data in the defined roots.

Authentication configuration is offered, allowing set up of token
authentication or TLS authentication to the Enterprise OPA instance as
described in the [Security
documentation](https://www.openpolicyagent.org/docs/latest/security/#authentication-and-authorization).
Tokens are set via command, so they are stored in VS Code secure storage
and not in configuration files.