Better output formatting and error handling (#71)
Previously when running previews, the output was a dump of JSON. This is
not always very friendly. This adapts functions from DAS UI to format
returns. In addition, extra data such as printed lines and
instrumentation metrics are formatted much more clearly. The response is
always at the (scrolled to) bottom so it is immediately visible.

In addition, non-200 API responses are handled and surfaced as an error
pop-up instead of printing JSON to the results output window. This make
it far easier to tell that something has gone wrong, and the EOPA
generated message is presented to the user along with the API response
code.