Initial reorganization and rewrite of docs (#73)
Since this project is moving beyond the Styra DAS integration with Styra
Link to also supporting features in Enterprise OPA, the documentation
and general branding needed a bit of a face lift. This renames the
plugin to 'Styra VS Code Tools', and reorganizes the README to allow
introducing the Enterprise OPA preview integration.

In addition, support has been consolidated into the generally supported
`SUPPORT.md` file, and the developer facing `README.md` has been
consolidated into the `CONTRIBUTING.md` file.

Lastly, in recognition that we want to fully embrace the name
'Enterprise OPA' for settings, the settings prefixes are updated to
render the name more fully and correctly.