styra-link: add check for bad url
With a bad URL the download still resulted in something,
(seemed to be a web page indicating 404) so the code
blithely installed that as /usr/local/bin/styra.
With that in place, trying to invoke it as an executable
resulted in the cryptic error
"Error spawn Unknown system error -8".
This change now recognizes that the non-empty download
was not an executable, so fails the "install" operation properly.